TFS + Question 5 answer
This commit contains my answer for question 5. For this question i only modified server code.

The files i modified are:
sebs_attack.lua,
spells.lua,
spells.xml

I have also posted links to the videos for setting up the environment as well as my video for question 5. The videos were large so i put links to them in the videos folder. If the shortcuts do not work you can find the links in the Videos/AllVideos.txt.

Thank you